Fix bootstrapper test and add mediator tests
