+++ FileProcessor.BusinessLogic/Common/GuidCalculator.cs
@@ -1,7 +1,9 @@
 ﻿namespace FileProcessor.BusinessLogic.Common
 {
     using System;
+    using System.Diagnostics.CodeAnalysis;
 
+    [ExcludeFromCodeCoverage]
     public static class GuidCalculator
     {
         #region Methods
