+++ FileProcessor.Testing/FileProcessor.Testing.csproj
@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <TargetFramework>net7.0</TargetFramework>
-	  <DebugType>Full</DebugType>
+	  <DebugType>None</DebugType>
   </PropertyGroup>
 
   <ItemGroup>
