+++ FileProcessor.Testing/TestData.cs
@@ -4,6 +4,7 @@ namespace FileProcessor.Testing
 {
     using System.Collections.Generic;
     using BusinessLogic.Managers;
+    using BusinessLogic.Requests;
     using EstateManagement.DataTransferObjects.Responses;
     using EstateReporting.Database.Entities;
     using File.DomainEvents;
@@ -787,5 +788,20 @@ namespace FileProcessor.Testing
                                                            },
 
                                                                      };
+        
+        public static UploadFileRequest UploadFileRequest =>
+            new UploadFileRequest(TestData.EstateId, TestData.MerchantId, TestData.UserId, TestData.FilePath, TestData.FileProfileId, TestData.FileUploadedDateTime);
+
+        public static ProcessUploadedFileRequest ProcessUploadedFileRequest =>
+            new ProcessUploadedFileRequest(TestData.EstateId, TestData.MerchantId, TestData.FileImportLogId, TestData.FileId, TestData.UserId, TestData.FilePath, TestData.FileProfileId,
+                                           TestData.FileUploadedDateTime);
+
+        public static SafaricomTopupRequest SafaricomTopupRequest => new SafaricomTopupRequest(TestData.FileId, TestData.FileName, TestData.FileProfileId);
+
+        public static VoucherRequest VoucherRequest => new VoucherRequest(TestData.FileId, TestData.FileName, TestData.FileProfileId);
+
+        public static ProcessTransactionForFileLineRequest ProcessTransactionForFileLineRequest =>
+            new ProcessTransactionForFileLineRequest(TestData.FileId, TestData.LineNumber, TestData.FileLine);
+
     }
 }
