+++ FileProcessor.Tests/FileProcessor.Tests.csproj
@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <TargetFramework>net7.0</TargetFramework>
-	  <DebugType>Full</DebugType>
+	  <DebugType>None</DebugType>
     <IsPackable>false</IsPackable>
   </PropertyGroup>
