+++ FileProcessor.sln
@@ -39,15 +39,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FileProcessor.Tests", "File
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FileProcessor.Client", "FileProcessor.Client\FileProcessor.Client.csproj", "{958AD704-4785-49B6-A90E-FEA2F63D92AC}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{9E93C30A-B086-41C9-ACDA-6B94BA769CD1}"
-	ProjectSection(SolutionItems) = preProject
-		.github\workflows\buildwindowsimage.yml = .github\workflows\buildwindowsimage.yml
-		.github\workflows\createrelease.yml = .github\workflows\createrelease.yml
-		.github\workflows\nightlybuild.yml = .github\workflows\nightlybuild.yml
-		.github\workflows\pullrequest.yml = .github\workflows\pullrequest.yml
-		.github\workflows\pushtomain.yml = .github\workflows\pushtomain.yml
-	EndProjectSection
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
