+++ FileProcessor/Bootstrapper/MiscRegistry.cs
@@ -1,13 +1,12 @@
 ﻿namespace FileProcessor.Bootstrapper;
 
 using BusinessLogic.Common;
+using BusinessLogic.Services;
 using Lamar;
 using Microsoft.Extensions.DependencyInjection;
+using System.Diagnostics.CodeAnalysis;
 
-/// <summary>
-/// 
-/// </summary>
-/// <seealso cref="Lamar.ServiceRegistry" />
+[ExcludeFromCodeCoverage]
 public class MiscRegistry : ServiceRegistry
 {
     #region Constructors
@@ -19,6 +18,7 @@ public class MiscRegistry : ServiceRegistry
     {
         this.AddSingleton<IModelFactory, ModelFactory>();
         this.AddSingleton<Common.IModelFactory, Common.ModelFactory>();
+        this.AddSingleton<IFileProcessorDomainService, FileProcessorDomainService>();
     }
 
     #endregion
