+++ FileProcessor/Startup.cs
@@ -10,14 +10,11 @@ using System.Threading.Tasks;
 
 namespace FileProcessor
 {
-    using System.Diagnostics;
     using System.Diagnostics.CodeAnalysis;
     using System.Globalization;
     using System.IO;
     using System.IO.Abstractions;
-    using System.Net.Http;
     using System.Reflection;
-    using System.Threading;
     using Bootstrapper;
     using BusinessLogic.Common;
     using BusinessLogic.EventHandling;
@@ -34,6 +31,7 @@ namespace FileProcessor
     using FIleProcessor.Models;
     using HealthChecks.UI.Client;
     using Lamar;
+    using LamarCodeGeneration.Util;
     using MediatR;
     using Microsoft.AspNetCore.Authentication.JwtBearer;
     using Microsoft.AspNetCore.Diagnostics.HealthChecks;
@@ -49,10 +47,8 @@ namespace FileProcessor
     using Shared.EntityFramework;
     using Shared.EntityFramework.ConnectionStringConfiguration;
     using Shared.EventStore.Aggregate;
-    using Shared.EventStore.EventHandling;
     using Shared.EventStore.EventStore;
     using Shared.EventStore.Extensions;
-    using Shared.EventStore.SubscriptionWorker;
     using Shared.Extensions;
     using Shared.General;
     using Shared.Logger;
@@ -147,7 +143,33 @@ namespace FileProcessor
 
             FileLineAddedEvent fileLineAddedEvent = new FileLineAddedEvent(Guid.Empty, Guid.Empty, 0, String.Empty);
 
-            TypeProvider.LoadDomainEventsTypeDynamically();
+            //TypeProvider.LoadDomainEventsTypeDynamically(new List<String>() {
+            //                                                                    "Pomelo",
+            //                                                                    "Microsoft"
+            //                                                                });
+
+            LoadDomainEventsTypeDynamically(new List<String>() {
+                                                                                "Pomelo",
+                                                                                "Microsoft"
+                                                                            });
+        }
+
+        public static void LoadDomainEventsTypeDynamically(List<string> assemblyFilters = null)
+        {
+            Assembly[] assemblies = AppDomain.CurrentDomain.GetAssemblies();
+            
+            List<Type> source = new List<Type>();
+            foreach (string assemblyFilter in assemblyFilters)
+            {
+                List<Assembly> filteredAssemblies = assemblies.Where(a => a.FullName.Contains(assemblyFilter) == true).ToList();
+                foreach (Assembly a in filteredAssemblies) {
+                    assemblies = assemblies.Remove(a);
+                }
+            }
+            source.AddRange(assemblies.SelectMany((Func<Assembly, IEnumerable<Type>>)(a => (IEnumerable<Type>)a.GetTypes())));
+
+            foreach (Type type in source.Where((Func<Type, bool>)(t => t.IsSubclassOf(typeof(DomainEvent)))).OrderBy((Func<Type, string>)(e => e.Name)).ToList())
+                TypeMap.AddType(type, type.Name);
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
@@ -205,153 +227,5 @@ namespace FileProcessor
 
             app.PreWarm();
         }
-    }
-
-    public static class Extensions
-    {
-        public static IServiceCollection AddInSecureEventStoreClient(
-            this IServiceCollection services,
-            Uri address,
-            Func<HttpMessageHandler>? createHttpMessageHandler = null)
-        {
-            return services.AddEventStoreClient((Action<EventStoreClientSettings>)(options =>
-                                                                                   {
-                                                                                       options.ConnectivitySettings.Address = address;
-                                                                                       options.ConnectivitySettings.Insecure = true;
-                                                                                       options.CreateHttpMessageHandler = createHttpMessageHandler;
-                                                                                   }));
-        }
-
-        static Action<TraceEventType, String, String> log = (tt, subType, message) => {
-            String logMessage = $"{subType} - {message}";
-            switch (tt)
-            {
-                case TraceEventType.Critical:
-                    Logger.LogCritical(new Exception(logMessage));
-                    break;
-                case TraceEventType.Error:
-                    Logger.LogError(new Exception(logMessage));
-                    break;
-                case TraceEventType.Warning:
-                    Logger.LogWarning(logMessage);
-                    break;
-                case TraceEventType.Information:
-                    Logger.LogInformation(logMessage);
-                    break;
-                case TraceEventType.Verbose:
-                    Logger.LogDebug(logMessage);
-                    break;
-            }
-        };
-
-        static Action<TraceEventType, String> mainLog = (tt, message) => Extensions.log(tt, "MAIN", message);
-        static Action<TraceEventType, String> orderedLog = (tt, message) => Extensions.log(tt, "ORDERED", message);
-
-        public static void PreWarm(this IApplicationBuilder applicationBuilder)
-        {
-            Startup.LoadTypes();
-
-            IConfigurationSection subscriptionConfigSection = Startup.Configuration.GetSection("AppSettings:SubscriptionConfiguration");
-            SubscriptionWorkersRoot subscriptionWorkersRoot = new SubscriptionWorkersRoot();
-            subscriptionConfigSection.Bind(subscriptionWorkersRoot);
-
-            if (subscriptionWorkersRoot.InternalSubscriptionService)
-            {
-                String eventStoreConnectionString = ConfigurationReader.GetValue("EventStoreSettings", "ConnectionString");
-
-                ISubscriptionRepository subscriptionRepository = SubscriptionRepository.Create(eventStoreConnectionString, subscriptionWorkersRoot.InternalSubscriptionServiceCacheDuration);
-                ((SubscriptionRepository)subscriptionRepository).Trace += (sender,
-                                                                           s) => Extensions.log(TraceEventType.Information, "REPOSITORY", s);
-
-                // init our SubscriptionRepository
-                subscriptionRepository.PreWarm(CancellationToken.None).Wait();
-
-                List<SubscriptionWorker> workers = ConfigureSubscriptions(subscriptionRepository, subscriptionWorkersRoot);
-                foreach (SubscriptionWorker subscriptionWorker in workers)
-                {
-                    subscriptionWorker.StartAsync(CancellationToken.None).Wait();
-                }
-            }
-        }
-
-        private static List<SubscriptionWorker> ConfigureSubscriptions(ISubscriptionRepository subscriptionRepository, SubscriptionWorkersRoot configuration)
-        {
-            List<SubscriptionWorker> workers = new List<SubscriptionWorker>();
-
-            foreach (SubscriptionWorkerConfig configurationSubscriptionWorker in configuration.SubscriptionWorkers)
-            {
-                if (configurationSubscriptionWorker.Enabled == false)
-                    continue;
-
-                if (configurationSubscriptionWorker.IsOrdered)
-                {
-                    IDomainEventHandlerResolver eventHandlerResolver = Startup.Container.GetInstance<IDomainEventHandlerResolver>("Ordered");
-                    SubscriptionWorker worker = SubscriptionWorker.CreateOrderedSubscriptionWorker(Startup.EventStoreClientSettings,
-                                                                                                   eventHandlerResolver,
-                                                                                                   subscriptionRepository,
-                                                                                                   configuration.PersistentSubscriptionPollingInSeconds);
-                    worker.Trace += (_,
-                                     args) => Extensions.orderedLog(TraceEventType.Information, args.Message);
-                    worker.Warning += (_,
-                                       args) => Extensions.orderedLog(TraceEventType.Warning, args.Message);
-                    worker.Error += (_,
-                                     args) => Extensions.orderedLog(TraceEventType.Error, args.Message);
-                    worker.SetIgnoreGroups(configurationSubscriptionWorker.IgnoreGroups);
-                    worker.SetIgnoreStreams(configurationSubscriptionWorker.IgnoreStreams);
-                    worker.SetIncludeGroups(configurationSubscriptionWorker.IncludeGroups);
-                    worker.SetIncludeStreams(configurationSubscriptionWorker.IncludeStreams);
-                    workers.Add(worker);
-                }
-                else
-                {
-                    for (Int32 i = 0; i < configurationSubscriptionWorker.InstanceCount; i++)
-                    {
-                        IDomainEventHandlerResolver eventHandlerResolver = Startup.Container.GetInstance<IDomainEventHandlerResolver>("Main");
-                        SubscriptionWorker worker = SubscriptionWorker.CreateSubscriptionWorker(Startup.EventStoreClientSettings,
-                                                                                                eventHandlerResolver,
-                                                                                                subscriptionRepository,
-                                                                                                configurationSubscriptionWorker.InflightMessages,
-                                                                                                configuration.PersistentSubscriptionPollingInSeconds);
-
-                        worker.Trace += (_,
-                                         args) => Extensions.mainLog(TraceEventType.Information, args.Message);
-                        worker.Warning += (_,
-                                           args) => Extensions.mainLog(TraceEventType.Warning, args.Message);
-                        worker.Error += (_,
-                                         args) => Extensions.mainLog(TraceEventType.Error, args.Message);
-
-                        worker.SetIgnoreGroups(configurationSubscriptionWorker.IgnoreGroups);
-                        worker.SetIgnoreStreams(configurationSubscriptionWorker.IgnoreStreams);
-                        worker.SetIncludeGroups(configurationSubscriptionWorker.IncludeGroups);
-                        worker.SetIncludeStreams(configurationSubscriptionWorker.IncludeStreams);
-
-                        workers.Add(worker);
-                    }
-                }
-            }
-
-            return workers;
-        }
-    }
-
-    public class SubscriptionWorkersRoot
-    {
-        public Boolean InternalSubscriptionService { get; set; }
-        public Int32 PersistentSubscriptionPollingInSeconds { get; set; }
-        public Int32 InternalSubscriptionServiceCacheDuration { get; set; }
-        public List<SubscriptionWorkerConfig> SubscriptionWorkers { get; set; }
-    }
-
-    public class SubscriptionWorkerConfig
-    {
-        public String WorkerName { get; set; }
-        public String IncludeGroups { get; set; }
-        public String IgnoreGroups { get; set; }
-        public String IncludeStreams { get; set; }
-        public String IgnoreStreams { get; set; }
-        public Boolean Enabled { get; set; }
-        public Int32 InflightMessages { get; set; }
-        public Int32 InstanceCount { get; set; }
-        public Boolean IsOrdered { get; set; }
     }
 }
